Converted Navbar to tsx
